# doc/swagger/swagger.yaml
...
swagger: '2.0'
info:
  title: argo-messaging-api
  description: "ARGO is a lightweight service for Service Level Monitoring designed for medium and large sized e-Infrastructures. This swagger UI interface provides access to the specification of the ARGO messaging api, which can be used to produce or consume messages."
  version: "0.9.2"
  contact:
    name: ARGO Developers
    url: http://argoeu.github.io/
    email: argo-dev@lists.grnet.gr
securityDefinitions:
  APIKeyHeader:
    type: apiKey
    in: header
    name: x-api-key
security:
  - APIKeyHeader: []
host: localhost
schemes:
  - https
basePath: /v1
produces:
  - application/json
...
paths:
...
  /registrations/{UUID}:
    get:
      summary: Retrieves information about a specific user registration
      description: |
        Retrieves a user's registration
      parameters:
        - $ref: '#/parameters/ApiKey'
        - name: UUID
          in: path
          description: UUID of the registration
          required: true
          type: string
      tags:
        - Registrations
      responses:
        200:
          description: User's registration
          schema:
            $ref: '#/definitions/UserRegistration'
        401:
          $ref: "#/responses/401"
        403:
          $ref: "#/responses/403"
        404:
          $ref: "#/responses/404"
        500:
          $ref: "#/responses/500"
...
  /status:
    get:
      summary: List the health status for ams and push server
      description: |
        list health status
      parameters:
        - $ref: '#/parameters/Details'
        - $ref: '#/parameters/ApiKeyDetails'
      tags:
        - Health Check
      responses:
        200:
          description: Returns the health status
          schema:
            $ref: '#/definitions/HealthCheck'
        401:
          $ref: "#/responses/401"
        403:
          $ref: "#/responses/403"
        500:
          $ref: "#/responses/500"
...
parameters:
  ApiKey:
    name: key
    in: query
    description: user key token for authentication
    required: true
    type: string
    default: SecretKey123
  ApiKeyDetails:
    name: key
    in: query
    description: user key token for authentication when accesing the url parameter details
    required: false
    type: string
    default: ""
  Details:
    name: details
    in: query
    description: report detailed error for ams push server
    required: false
    type: boolean
    default: false
...
responses:
...
  401:
    description: Unauthorized user based on key
    schema:
      $ref: '#/definitions/ErrorMsg'
  403:
    description: Access Forbidden for the user on the resource
    schema:
      $ref: '#/definitions/ErrorMsg'
  404:
    description: Item not found
    schema:
      $ref: '#/definitions/ErrorMsg'
...
  500:
    description: Internal Error
    schema:
      $ref: '#/definitions/ErrorMsg'
definitions:
  UserRegistration:
    type: object
    properties:
      uuid:
        type: string
      name:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      organization:
        type: string
      description:
        type: string
      activation_token:
        type: string
      status:
        type: string
      registered_at:
        type: string
      modified_at:
        type: string
      modified_by:
        type: string
...
  HealthCheck:
    type: object
    properties:
      status:
        type: string
      push_servers:
        type: array
        items:
          $ref: '#/definitions/PushServerInfo'
  PushServerInfo:
    type: object
    properties:
      endpoint:
        type: string
      status:
        type: string
...
  ErrorMsg:
    type: object
    properties:
      error:
        type: object
        properties:
          code:
            type: string
            description: code of the error
          message:
            type: string
            description: general message of the error
          status:
            type: string
            description: status of the error
...